requirements.txt
pyjwt>=2.6.0
colorama>=0.4.4
pyperclip>=1.8.2
cryptography>=39.0.0 